+++ packages/ebsr/configure/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.10](https://github.com/pie-framework/pie-elements/compare/@pie-element/ebsr-configure@1.2.9...@pie-element/ebsr-configure@1.2.10) (2019-05-07)
+
+**Note:** Version bump only for package @pie-element/ebsr-configure
+
+
+
+
+
+## [1.2.9](https://github.com/pie-framework/pie-elements/compare/@pie-element/ebsr-configure@1.2.8...@pie-element/ebsr-configure@1.2.9) (2019-05-02)
+
+**Note:** Version bump only for package @pie-element/ebsr-configure
+
+
+
+
+
 ## [1.2.8](https://github.com/pie-framework/pie-elements/compare/@pie-element/ebsr-configure@1.2.7...@pie-element/ebsr-configure@1.2.8) (2019-05-02)
 
 **Note:** Version bump only for package @pie-element/ebsr-configure