+++ packages/extended-text-entry/configure/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry-configure@2.4.4...@pie-element/extended-text-entry-configure@2.4.5) (2019-05-02)
+
+**Note:** Version bump only for package @pie-element/extended-text-entry-configure
+
+
+
+
+
 ## [2.4.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry-configure@2.4.3...@pie-element/extended-text-entry-configure@2.4.4) (2019-05-02)
 
 **Note:** Version bump only for package @pie-element/extended-text-entry-configure